pyproject.toml
[project]
name = "ms-with-aiogram"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.20.0.post0",
    "faststream[rabbit]>=0.5.44",
    "python-dotenv>=1.1.1",
]
